# perf: projective point for ecdsa ops