Migrated all multiplications between EC points and field elements in ECDSA to projective points. Included a new benchmark for the `ecdsa_recover` function.
Refactored the `ecdsa_verify` changes too so all use a little helper to reduce boilerplate.

Results:
```
ecdsa_get_public_key    time:   [113.13 µs 113.20 µs 113.28 µs]
                        change: [-92.975% -92.959% -92.940%] (p = 0.00 < 0.05)
                        Performance has improved.

ecdsa_recover           time:   [412.73 µs 413.33 µs 414.48 µs]
                        change: [-88.145% -88.120% -88.082%] (p = 0.00 < 0.05)
                        Performance has improved.

ecdsa_sign              time:   [160.00 µs 160.06 µs 160.14 µs]
                        change: [-90.003% -89.985% -89.967%] (p = 0.00 < 0.05)
                        Performance has improved.

ecdsa_verify            time:   [412.61 µs 412.77 µs 412.98 µs]
                        change: [-3.8553% -1.8908% -0.5588%] (p = 0.01 < 0.05)
                        Change within noise threshold.
```